fix: properly handle expiring DDNS records
Some DDNS providers, such as dy.fi, have a 7-day expiration for DDNS records if they're not updated, even if the associated IP hasn't changed. ddclient isn't smart enough to figure that out, so remove its cache every 5 days to force it to perform an update and keep the record alive.

Signed-off-by: Dennis Marttinen <twelho@welho.tech>